fix: fixed some beauty mistakes in the code.

+++ comsat/src/main/java/nl/codefusion/comsat/controller/ContactController.java
@@ -5,6 +5,7 @@ import nl.codefusion.comsat.config.Permission;
 import nl.codefusion.comsat.dao.ContactDao;
 import nl.codefusion.comsat.models.ContactModel;
 import nl.codefusion.comsat.repository.ContactRepository;
+import nl.codefusion.comsat.service.ContactService;
 import nl.codefusion.comsat.service.PermissionService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -24,35 +25,32 @@ public class ContactController {
 
     private final ContactDao contactDao;
 
+    private final ContactService contactService;
+
     @PostMapping("/contacts")
-    public ContactModel createContact(@RequestBody ContactModel contact) {
-        return contactRepository.save(contact);
+    public ContactModel createContact(@RequestBody ContactModel contact) throws NoPermissionException {
+        if (permissionService.hasPermission(permissionService.getPrincipalRoles(), Permission.CREATE_CONTACT)) {
+            return contactRepository.save(contact);
+        }
+        throw new NoPermissionException();
     }
 
     @DeleteMapping("/contacts/{id}")
-    public ResponseEntity<Void> deleteContact(@PathVariable UUID id) {
-        contactRepository.deleteById(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<Void> deleteContact(@PathVariable UUID id) throws NoPermissionException {
+        if (permissionService.hasPermission(permissionService.getPrincipalRoles(), Permission.READ_CONTACT_DETAILS)) {
+            contactRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+        throw new NoPermissionException();
+
     }
 
     @PutMapping("/contacts/{id}")
-    public ResponseEntity<ContactModel> updateContact(@PathVariable UUID id, @RequestBody ContactModel contactDetails) {
-        ContactModel contact = contactRepository.findById(id).orElseThrow();
-
-        if (contactDetails.getFirstName() != null) {
-            contact.setFirstName(contactDetails.getFirstName());
+    public ResponseEntity<ContactModel> updateContact(@PathVariable UUID id, @RequestBody ContactModel contactDetails) throws NoPermissionException {
+        if (permissionService.hasPermission(permissionService.getPrincipalRoles(), Permission.UPDATE_CONTACT)) {
+            return contactService.updateContact(id, contactDetails);
         }
-        if (contactDetails.getNickname() != null) {
-            contact.setNickname(contactDetails.getNickname());
-        }
-        if (contactDetails.getAudience() != null) {
-            contact.setAudience(contactDetails.getAudience());
-        }
-        if (contactDetails.getSex() != null) {
-            contact.setSex(contactDetails.getSex());
-        }
-        ContactModel updatedContact = contactRepository.save(contact);
-        return ResponseEntity.ok(updatedContact);
+        throw new NoPermissionException();
     }
 
     @GetMapping
